Mr President... they've gotten the second floor
second floor in, very messy code, not proud.
only 1 entrance / exit to second floor area.
on main floor generates a specific 2 high prefab which is "transitional" and acts as starting point of second floor maze.
transitional prefab has an "exit" marked as one of its "1x1" cells, thats the acc starting point of the maze generation.
it works, not proud, but it works.